Rework Puppet provisioner for bind mounts
The Puppet provisioner used to copy a single manifest into
the rootfs of an LXC and just apply it with a fixed module
directory.

When using bind mounts, one can easily mount a manifests and
module directory into the LXC. This commit changes the puppet
provisioner to take absolute paths to an initial manifest and
the modules directory in the container.